+++ src/Pim/Bundle/InstallerBundle/DependencyInjection/PimInstallerExtension.php
@@ -22,6 +22,7 @@ class PimInstallerExtension extends Extension
         $loader = new Loader\YamlFileLoader($container, new FileLocator(__DIR__.'/../Resources/config'));
         $loader->load('guessers.yml');
         $loader->load('transformers.yml');
+        $loader->load('fixture_loader.yml');
         $this->addInstallerDataFiles($container);
     }
 
@@ -36,7 +37,7 @@ class PimInstallerExtension extends Extension
         preg_match('/^(?P<bundle>\w+):(?P<directory>\w+)$/', $dataParam, $matches);
         $bundles    = $container->getParameter('kernel.bundles');
         $reflection = new \ReflectionClass($bundles[$matches['bundle']]);
-        $dataPath   = dirname($reflection->getFilename()).'/Resources/config/installer/'.$matches['directory'].'/';
+        $dataPath   = dirname($reflection->getFilename()) . '/Resources/fixtures/' . $matches['directory'] . '/';
 
         $entities = array(
             'channels',
@@ -55,9 +56,12 @@ class PimInstallerExtension extends Extension
         $installerFiles = array();
 
         foreach ($entities as $entity) {
-            $file = $dataPath.$entity.'.yml';
-            if (is_file($file)) {
-                $installerFiles[$entity]= $file;
+            $file = $dataPath.$entity;
+            foreach (array('.yml', '.csv') as $extension) {
+                if (is_file($file . $extension)) {
+                    $installerFiles[$entity] = $file . $extension;
+                    break;
+                }
             }
         }
         $container->setParameter('pim_installer.files', $installerFiles);