+++ src/Pim/Bundle/UIBundle/Resources/public/js/pim-init.js
@@ -1,7 +1,7 @@
 define(
-    ['jquery', 'oro/translator', 'oro/mediator', 'oro/navigation', 'oro/messenger', 'pim/dialog', 'oro/loading-mask',
-     'pim/initselect2', 'pim/saveformstate', 'bootstrap', 'bootstrap.bootstrapswitch', 'bootstrap-tooltip'],
-    function ($, __, mediator, Navigation, messenger, Dialog, LoadingMask, initSelect2, saveformstate) {
+    ['jquery', 'oro/translator', 'oro/mediator', 'oro/navigation', 'oro/messenger', 'pim/dialog',
+     'pim/saveformstate', 'pim/asynctab', 'pim/ui'],
+    function ($, __, mediator, Navigation, messenger, Dialog, saveformstate, loadTab, UI) {
         'use strict';
         var initialized = false;
         return function() {
@@ -9,30 +9,15 @@ define(
                 return;
             }
             initialized = true;
-            function loadTab(tab) {
-                var $tab = $(tab);
-                var target = $tab.attr('href');
-                if (!target || target === '#' || target.indexOf('javascript') === 0) {
-                    return;
+            var setFullHeight = function ($target) {
+                if (!$target) {
+                    $target = $('body');
                 }
-                var $target = $(target);
-
-                if (!$target.attr('data-loaded') && $target.attr('data-url')) {
-                    var loadingMask = new LoadingMask();
-                    loadingMask.render().$el.appendTo($('#container'));
-                    loadingMask.show();
-
-                    $.get($target.attr('data-url'), function(data) {
-                        $target.html(data);
-                        $target.attr('data-loaded', 1);
-                        loadingMask.hide();
-                        loadingMask.$el.remove();
-                        $target.closest('form').trigger('tab.loaded');
-                        pageInit($target);
-                    });
-                }
-            }
-            function pageInit($target) {
+                $target.find('.fullheight').filter(':visible').each(function () {
+                    $(this).height($('.scrollable-container').height() - $(this).position().top + $('.scrollable-container').position().top);
+                });
+            };
+            var pageInit = function ($target) {
                 if (!$target) {
                     $target = $('body');
                     $target.find('form.form-horizontal').each(function() {
@@ -46,24 +31,6 @@ define(
                     if (target.length) {
                         $(this).appendTo(target).attr('tabIndex', -1);
                     }
-                });
-
-                // Apply Select2
-                initSelect2.init($target);
-
-                // Apply bootstrapSwitch
-                $target.find('.switch:not(.has-switch)').bootstrapSwitch();
-
-                // Initialize tooltip
-                $target.find('[data-toggle="tooltip"]').tooltip();
-
-                // Initialize popover
-                $target.find('[data-toggle="popover"]').popover();
-
-                // Activate a form tab
-                $target.find('li.tab.active a').each(function () {
-                    var paneId = $(this).attr('href');
-                    $(paneId).addClass('active');
                 });
 
                 // Toogle accordion icon
@@ -80,8 +47,10 @@ define(
                     }
                 });
                 $target.find('.attribute-field.translatable').each(function () {
-                    $(this).find('div.controls').find('.icons-container').eq(0).prepend($localizableIcon.clone().tooltip());
+                    $(this).find('div.controls').find('.icons-container').eq(0).prepend($localizableIcon.clone());
                 });
+
+                UI($target);
 
                 $target.find('a[data-form-toggle]').on('click', function () {
                     $('#' + $(this).attr('data-form-toggle')).show();
@@ -91,7 +60,9 @@ define(
                 $target.find('a[data-toggle="tab"]').on('show.bs.tab', function() {
                     loadTab(this);
                 });
-            }
+
+                setFullHeight($target);
+            };
 
             $(function(){
                 if ($.isPlainObject($.uniform)) {
@@ -100,6 +71,17 @@ define(
 
                 $(document).on('uniformInit', function () {
                     $.uniform.restore();
+                });
+
+                $(document).on('tab.loaded', 'form.form-horizontal', function(e, tab) {
+                    pageInit($(tab));
+                });
+
+                $(document).on('shown', 'a[data-toggle="tab"]', function() {
+                    var target = $(this).attr('href');
+                    if (target && target !== '#' && target.indexOf('javascript') !== 0) {
+                        setFullHeight($(target).parent());
+                    }
                 });
 
                 // DELETE request for delete buttons
@@ -111,7 +93,7 @@ define(
                             $.ajax({
                                 url: $el.attr('data-url'),
                                 type: 'POST',
-                                headers: {accept:'application/json'},
+                                headers: { accept:'application/json' },
                                 data: { _method: $el.data('method') },
                                 success: function() {
                                     var navigation = Navigation.getInstance();
@@ -128,7 +110,7 @@ define(
                             });
                         };
                     $el.off('click');
-                    if ($el.data('dialog') ===  'confirm') {
+                    if ($el.data('dialog') === 'confirm') {
                         Dialog.confirm(message, title, doAction);
                     } else {
                         Dialog.alert(message, title);